+++ open360tracker/open360tracker.ino
@@ -460,7 +460,7 @@ void loop()
         trackerPosition.lon = trackerPosition.lon / 10;
 
         if (gps.altitude() != TinyGPS::GPS_INVALID_ALTITUDE) {
-          trackerPosition.alt = int16_t(gps.altitude());
+          trackerPosition.alt = int32_t(gps.altitude());
         }
 
         if (gps.satellites() != TinyGPS::GPS_INVALID_SATELLITES){
